saturday morning

+++ src/Util.java
@@ -17,8 +17,8 @@ public class Util {
     }
 
     // Euclidean distance
-    static float edist(int x1, int y1, int x2, int y2){
-        return (float)Math.sqrt(Math.pow(x1-x2,2) + Math.pow(y1-y2,2));
+    static double edist(int x1, int y1, int x2, int y2){
+        return Math.sqrt(Math.pow(x1-x2,2) + Math.pow(y1-y2,2));
     }
 
     static public <T> T[] concatenate(T[] a, T[] b) {
